Big change to lock.h
1. There are now two clear divisions to the file. One section that defines
a set of macros/functions if DEBUG_THREADS is defined, and one that defines the same
set if DEBUG_THREADS is not defined. This change paved the way for...

2. Handling the storage of backtraces for read-write locks was tricky since the
ast_rwlock_t was just a typedef of pthread_rwlock_t. Now, if DEBUG_THREADS is
defined, ast_rwlock_t is a structure that contains a lock and a backtrace. This
allows for easy removal of the backtrace associated with the lock when it is unlocked
(although I admit now that I have not done that yet. Next commit!!)

3. Because of the two above changes, there is no dynamic allocation of ast_bt's 
any more, so I have removed the ast_bt_destroy calls from utils.c.