Check in a work in progress.  While half awake one night I got to thinking about media handling in chan_iax2.  The way frames get sent up to the core from the IAX2 processing threads is by queueing frames onto the ast_channel. However, it's very expensive as it requires malloc'ing a new frame and expensive locking with deadlock avoidance (accessing an ast_channel from a channel driver is usually pretty ugly, as Asterisk locking rules require that you get the ast_channel lock _before_ the channel's pvt lock, and coming from the channel driver, you usually have the pvt before the channel.)
So, anyway, I started playing around with ways to pass audio to the core by
queueing the audio on to the pvt struct, and allowing the core to use the
iax2_read callback to get it.  So far, I have avoided the nasty deadlock
avoidance that had to be done when queueing frames, and it appears to have
some performance improvements already.  However, I still have more I want to
do to reduce how many times frames are malloc'd and copied around within
chan_iax2 ...