more debugging for memory leaks (soon to go away).
It appears that frames on chan->readq are not freed
properly if there is no receiver on the remote side.

Because video console still tries to encode and enqueue
frames even when a connection is not active, the queue is
drained and apparently ast_frames are thrown away without
freeing.