Some new events:
CEL_PICKUP - report a directed pickup event so you know what happens
when extension 200 is called but extension 300 arrives at the bridge.

CEL_FORWARD - signal a 302 redirect or other protocol-level forwarding
so that you can tell why a call runs off in a different direction due
to a phone being set to forward calls.

CEL_BRIDGE_UPDATE - report when a participant is masqueraded into a
bridge.  My primary use for this is to figure out what's going on when
a bridge occurs with a chan_local channel which is then masqueraded
out when the remote party answers.