Merged revisions 127297 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r127297 | tilghman | 2008-07-01 21:48:43 -0500 (Tue, 01 Jul 2008) | 12 lines

Change the global timer B to be dependent on the value of the T1 timer, as
recommended in RFC 3261, instead of being hardcoded to 32 seconds.  This is
important for LANs, as it allows autocongestion to occur much more quickly, if
desired by the local PBX administrator.  It also corrects a bug: if the T1
timer was increased beyond 500ms, then timer B would have been set at a much
lower value than recommended.
(closes issue #12544)
 Reported by: kactus
 Patches: 
       20080616__bug12544.diff.txt uploaded by Corydon76 (license 14)
 Tested by: kactus

........