Based on the problem reported in issue #12764, I did some investigation of app_voicemail and have found massive inconsistencies when using IMAP storage and using other types of storage.
When using ODBC storage, if a file needs to be manipulated, the file is downloaded
and placed into temporary storage on the file system along with a text file with
metadata. This way, the message can be manipulated in the same way as if file storage
were used.

IMAP, on the other hand, never actually creates a text file and seems to be inconsistent
with when the sound file is downloaded and destroyed.

The goal of this branch is to make IMAP's RETRIEVE and DISPOSE macros mimick the
ODBC implementation. By taking these simple steps, I believe that several of the
IMAP-only routines in app_voicemail can be completely removed, thus reducing the
code size and making app_voicemail more consistent.

I've created this branch against the 1.4 Asterisk branch because in my view, all of
this work is a huge bug fix.