Merged revisions 131570,131606 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r131570 | murf | 2008-07-16 17:53:02 -0600 (Wed, 16 Jul 2008) | 20 lines

(closes issue #13089)
Reported by: murf

Most of this bug was already fixed by Tilghman before
I opened it; Many thanks to Tilghman for his fix
in svn version 125794. That fix cleared up some of the
fields in the lock_info.

This commit changes the address that is stored for the
lock in the lock_info struct, so that it is the same 
as that passed into the locking macros. This makes 
searching for a lock_info (as in log_show_lock()) 
by its lock addr possible. The lock_addr field is
infinitely more useful if it is the same as what
is 'publicly' available outside the lock_info code.

Many thanks to kpfleming, putnopvut, and Russell for their
invaluable insights earlier today.


........
r131606 | tilghman | 2008-07-17 08:00:27 -0600 (Thu, 17 Jul 2008) | 7 lines

Change several 'core' commands to be 'dialplan' commands (with appropriate
deprecation, of course)
(closes issue #13016)
 Reported by: caio1982
 Patches: 
       dialplan_globals6.diff uploaded by caio1982 (license 22)

........